+++ db/migrations/000003_create_phrases_table.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS phrases (
 
 CREATE TABLE IF NOT EXISTS translates (
     phrase_id bigint NOT NULL REFERENCES phrases ON DELETE CASCADE,
-    language_id int NOT NULL REFERENCES languages ON DELETE CASCADE,
+    language_id smallint NOT NULL REFERENCES languages ON DELETE CASCADE,
     phrase text NOT NULL,
     phrase_hint text NOT NULL,
     PRIMARY KEY (language_id, phrase_id)
